+++ src/adc/class_lister.py
@@ -24,4 +24,7 @@ def list_classes() -> Dict[str, List[str]]:
         "kasperl.api.Generator": [
             "kasperl.generator",
         ],
+        "kasperl.api.DataFormatter": [
+            "adc.dataformatter",
+        ],
     }