+++ src/adc/dataformatter/__init__.py
@@ -1,0 +1,1 @@
+from ._audio_data_formatter import AudioDataFormatter